third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/12.14.1-expected.txt

PASS Function("'use strict'; try{}catch(eval){}") threw exception of type SyntaxError.
PASS !!Function("try{}catch(eval){}") is true
PASS true === true
PASS Function("'use strict'; try{}catch([eval]){}") threw exception of type SyntaxError.
PASS !!Function("try{}catch([eval]){}") is true
PASS true === true
PASS Function("'use strict'; try{}catch({x:eval}){}") threw exception of type SyntaxError.
PASS !!Function("try{}catch({x:eval}){}") is true
PASS true === true
PASS Function("'use strict'; try{}catch(arguments){}") threw exception of type SyntaxError.
PASS !!Function("try{}catch(arguments){}") is true
PASS true === true
PASS Function("'use strict'; try{}catch([arguments]){}") threw exception of type SyntaxError.
PASS !!Function("try{}catch([arguments]){}") is true
PASS true === true
PASS Function("'use strict'; try{}catch({x:arguments}){}") threw exception of type SyntaxError.
PASS !!Function("try{}catch({x:arguments}){}") is true
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
